Move MapForm helpers in own file.
